Removed extra package-info, removed old dependency, moved config outside of impl
