+++ sdk/appconfiguration/spring-cloud-azure-appconfiguration-config-web/src/main/java/com/azure/spring/cloud/config/web/AppConfigurationWebAutoConfiguration.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.spring.cloud.config.web.implementation.config;
+package com.azure.spring.cloud.config.web;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
