--- sdk/appconfiguration/spring-cloud-azure-appconfiguration-config-web/src/main/java/com/azure/spring/cloud/config/web/implementation/config/package-info.java
@@ -1,1 +1,0 @@
-package com.azure.spring.cloud.config.web.implementation.config;
