+++ sdk/appconfiguration/spring-cloud-azure-appconfiguration-config-web/src/main/java/com/azure/spring/cloud/config/web/package-info.java
@@ -1,0 +1,1 @@
+package com.azure.spring.cloud.config.web;
