--- sdk/appconfiguration/spring-cloud-azure-appconfiguration-config/src/main/java/com/azure/spring/cloud/config/implementation/feature/management/entity/package-info.java
@@ -1,1 +1,0 @@
-package com.azure.spring.cloud.config.implementation.feature.management.entity;
