--- sdk/appconfiguration/spring-cloud-azure-appconfiguration-config/src/main/java/com/azure/spring/cloud/config/implementation/pipline/policies/package-info.java
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-/**
- * Package contains classes for properties used to configure the library.
- */
-package com.azure.spring.cloud.config.implementation.pipline.policies;
