--- sdk/appconfiguration/spring-cloud-azure-appconfiguration-config/src/main/java/com/azure/spring/cloud/config/implementation/stores/package-info.java
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-/**
- * Package contains classes for using the Azure SDK for Java.
- */
-package com.azure.spring.cloud.config.implementation.stores;
